requirements.txt
Flask==3.1.3
Flask_Migrate==4.1.0
flask_wtf==1.2.2
injector==0.24.0
openai==2.31.0
pytest==9.0.2
python-dotenv==1.2.2
WTForms==3.2.1
